requirements.txt
streamlit
pandas
pillow
gspread
google-auth
